reduce polling, use instance IDs references
Use instance id as a reference to instance in environment to make sure
we are always operating on the same object that instance service has.

Reduce polling of instances to 60 seconds if no instances in transient
states are detected.

Clear interval in environment service on destroy.

Simplify update logic of referenced instances in environment service,
also make sure we do initial loading of data in correct order.

Bonus: fix workspace member listing and non-existing user query for
       accessible workspaces in mock interceptor