deal with failed instances
Environment launch failure (a failed instance in the backend) can now be
handled by the user. Desktop notification service can also track the
notifications for failures, so it does not keep on spamming. There is an
option in the mock interceptor to mimic random failures. Restarting
the environment is not straight forward as it will include state
tracking (ask backend to delete the instance, wait for the deletion to
complete, start a new one), so failed instances have to be deleted
first.