add "Always pull image" to application config
Sometimes customers reuse the same image tag for updated images as
well. In this case, we need to be able to force checking if the image
on the node is actually up-to-date. Cache will still be honored, so even
if we pull we don't transfer existing layers again.

Also drop custom image from the Wizard as it is potentially confusing.
Advanced use cases can be done in Edit.